Add toggle functionality for the Performance UI
In the diagnostic.rs file, a new function 'toggle' is added to create or despawn the Performance UI upon pressing KeyCode::F12. This also includes ordering the system before PerfUiSet::Setup to process new Perf UI in the same frame as entity spawning. The diagnostic module is also made optional depending on the wasm32 target architecture. Additional index.html file was added for the WASM example and edits on 'resources.md' was made to include a guide on building the WASM target.